RFC: utils: create a separate module
