+++ dvc/command/imp.py
@@ -1,7 +1,7 @@
 import os
 
 import dvc.logger as logger
-from dvc.utils import urlparse
+from dvc.utils.compat import urlparse
 from dvc.exceptions import DvcException
 from dvc.command.base import CmdBase
